+++ DP/Lib/GridWorldPrinter.h
@@ -11,7 +11,7 @@
 namespace GridWorld {
 
 void PrintValues(DP::ValueFunction &value_function, const int grid_size) {
-  float *const values = value_function.Values();
+  double *const values = value_function.Values();
   for (int i = 0; i < grid_size; ++i) {
     for (int j = 0; j < grid_size; ++j)
       std::cout << std::setprecision(3) << std::setw(6)
